% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{countlines}
\alias{countlines}
\title{Count the number of lines in a file}
\usage{
countlines(fn)
}
\arguments{
\item{fn}{Input filepath}
}
\value{
A integer for the number of lines
}
\description{
Count the number of lines in a file
}

