% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readbed.R
\name{getJArray}
\alias{getJArray}
\title{Import Java array into R}
\usage{
getJArray(mat_ref, na_vals = -9)
}
\arguments{
\item{mat_ref}{Reference object of the Java array}

\item{na_vals}{NA code. Default to -9.}
}
\description{
A thin wrapper around \code{rJava::.jevalArray}
}
\author{
Kaiyin Zhong
}

