% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ncols}
\alias{ncols}
\title{Counts the number of columns of whitespace delimited file.}
\usage{
ncols(fn)
}
\arguments{
\item{fn}{Input filepath}
}
\value{
A integer for the number of columns
}
\description{
Counts the number of columns of whitespace delimited file.
}

