% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{readcol}
\alias{readcol}
\title{Read one column of a whitespace delimited text file}
\usage{
readcol(fileName, colNum, nSkip, maxRowNum)
}
\arguments{
\item{fileName}{Input filepath}

\item{colNum}{An integer. The target column number}

\item{nSkip}{An integer. Number of lines to skip in the beginning.}

\item{maxRowNum}{An Integer. Maximum number of lines to read}
}
\value{
A vector of strings containing the target column
}
\description{
Read one column of a whitespace delimited text file
}

