% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{readcols}
\alias{readcols}
\title{Read columns of a whitespace delimited text file}
\usage{
readcols(fn, colsel, nFirstSkipLines, nSkipUnit)
}
\arguments{
\item{fn}{input filepath}

\item{colsel}{a vector of target column numbers}

\item{nFirstSkipLines}{Integer. Number of lines to skip in the beginning}

\item{nSkipUnit}{Integer M. Let the function read one line out of every M}
}
\value{
A matrix of strings from selected columns
}
\description{
Read columns of a whitespace delimited text file
}

