% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readbed.R
\name{truncateEndOfFile}
\alias{truncateEndOfFile}
\title{Truncate n bytes from end of file}
\usage{
truncateEndOfFile(filename, len)
}
\arguments{
\item{filename}{character. Filename.}

\item{len}{numeric. Number of bytes to truncate}
}
\description{
Truncate n bytes from end of file
}
\examples{
\dontrun{
fn = tempfile()
f = file(fn, "wb")
writeBin("a", f)
writeBin("b", f)
writeBin("c", f)
close(f)
file.info(fn)$size == 6
truncateEndOfFile(fn, 1)
file.info(fn)$size == 5
}
}
\author{
Kaiyin Zhong
}

